requirements.txt
# Assignment 3: Sequential Instruction Tuning
# Python >= 3.10 required
# Install: pip install -r requirements.txt
# On ARC:  pip install -r requirements.txt --extra-index-url https://download.pytorch.org/whl/cu118

# ── Core deep learning ────────────────────────────────────────
torch>=2.1.0
torchvision>=0.16.0

# ── HuggingFace ecosystem ─────────────────────────────────────
transformers>=4.44.0
datasets>=2.20.0
peft>=0.12.0
trl>=0.9.0
accelerate>=0.33.0
huggingface-hub>=0.24.0
tokenizers>=0.19.0

# ── Quantization ──────────────────────────────────────────────
bitsandbytes>=0.43.1

# ── Evaluation metrics ────────────────────────────────────────
rouge-score>=0.1.2
bert-score>=0.3.13
evaluate>=0.4.2

# ── Data / utilities ──────────────────────────────────────────
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0.1
tqdm>=4.66.0
scikit-learn>=1.3.0

# ── Logging (optional — comment out if not using W&B) ─────────
# wandb>=0.17.0

# ── Visualization ─────────────────────────────────────────────
matplotlib>=3.7.0
seaborn>=0.12.0
